Add test report dir to native
