+++ kotest-framework/kotest-framework-plugin-gradle/api/kotest-framework-plugin-gradle.api
@@ -1,6 +1,7 @@
 public abstract class io/kotest/framework/gradle/KotestPlugin : org/gradle/api/Plugin {
 	public static final field Companion Lio/kotest/framework/gradle/KotestPlugin$Companion;
 	public static final field TASK_DESCRIPTION Ljava/lang/String;
+	public static final field TESTS_DIR_NAME Ljava/lang/String;
 	public fun <init> ()V
 	public synthetic fun apply (Ljava/lang/Object;)V
 	public fun apply (Lorg/gradle/api/Project;)V
@@ -15,16 +16,11 @@ public final class io/kotest/framework/gradle/SpecsResolver {
 }
 
 public abstract class io/kotest/framework/gradle/tasks/AbstractKotestTask : org/gradle/api/DefaultTask {
-	public static final field Companion Lio/kotest/framework/gradle/tasks/AbstractKotestTask$Companion;
-	public static final field TESTS_DIR_NAME Ljava/lang/String;
 	public abstract fun getDescriptor ()Lorg/gradle/api/provider/Property;
 	public abstract fun getPackages ()Lorg/gradle/api/provider/Property;
 	public abstract fun getSpecs ()Lorg/gradle/api/provider/Property;
 	public abstract fun getTags ()Lorg/gradle/api/provider/Property;
-	protected final fun getTestReportsDir (Lorg/gradle/api/Project;Ljava/lang/String;)Lorg/gradle/api/provider/Provider;
-}
-
-public final class io/kotest/framework/gradle/tasks/AbstractKotestTask$Companion {
+	public abstract fun getTestReportsDir ()Lorg/gradle/api/provider/Property;
 }
 
 public abstract class io/kotest/framework/gradle/tasks/KotestAndroidTask : io/kotest/framework/gradle/tasks/AbstractKotestTask {
@@ -56,6 +52,7 @@ public abstract class io/kotest/framework/gradle/tasks/KotestJvmTask : io/kotest
 
 public abstract class io/kotest/framework/gradle/tasks/KotestNativeTask : io/kotest/framework/gradle/tasks/AbstractKotestTask {
 	protected final fun execute ()V
+	public abstract fun getExe ()Lorg/gradle/api/provider/Property;
 	public abstract fun getTarget ()Lorg/gradle/api/provider/Property;
 }
