+++ kotest-framework/kotest-framework-plugin-gradle/src/main/kotlin/io/kotest/framework/gradle/tasks/KotestJvmTask.kt
@@ -2,10 +2,8 @@ package io.kotest.framework.gradle.tasks
 
 import io.kotest.framework.gradle.SpecsResolver
 import io.kotest.framework.gradle.TestLauncherJavaExecConfiguration
-import org.gradle.api.file.Directory
 import org.gradle.api.plugins.JavaPluginExtension
 import org.gradle.api.provider.Property
-import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.CacheableTask
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.StopExecutionException
@@ -28,7 +26,6 @@ abstract class KotestJvmTask @Inject internal constructor(
    protected fun execute() {
 
       val java = project.extensions.getByType(JavaPluginExtension::class.java)
-      val testReportsDir: Provider<Directory> = getTestReportsDir(project, name)
 
       val test = java.sourceSets.findByName(sourceSetName.get())
          ?: throw StopExecutionException("Could not find source set '${sourceSetName.get()}'")
